LayoutTests/fast/text/variations/advances.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<div><span id="regular" style="font: 100px 'Skia';">ham</span></div>
<div><span id="wide" style="font: 100px 'Skia'; font-variation-settings: 'wdth' 1.25;">ham</span></div>
<div><span id="specifiedregular" style="font: 100px 'Skia'; font-variation-settings: 'wdth' 1;">ham</span></div>
<div><span id="narrow" style="font: 100px 'Skia'; font-variation-settings: 'wdth' 0.62;">ham</span></div>
<script>
var regular = document.getElementById("regular");
var wide = document.getElementById("wide");
var specifiedregular = document.getElementById("specifiedregular");
var narrow = document.getElementById("narrow");

shouldNotBe("regular.offsetWidth", "wide.offsetWidth");
shouldNotBe("regular.offsetWidth", "narrow.offsetWidth");
shouldNotBe("narrow.offsetWidth", "wide.offsetWidth");
shouldBe("regular.offsetWidth", "specifiedregular.offsetWidth");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
